+++ Sora/PeerChannel.swift
@@ -47,6 +47,9 @@ final class PeerChannelInternalHandlers {
   /// シグナリング受信時に呼ばれるクロージャー
   var onReceiveSignaling: ((Signaling) -> Void)?
 
+  /// シグナリング受信時に JSON 文字列で呼ばれるクロージャー
+  var onReceiveSignalingJSON: ((String) -> Void)?
+
   /// DataChannel の open 時に呼ばれるクロージャー
   var onOpenDataChannel: ((String) -> Void)?
 
@@ -182,6 +185,10 @@ class PeerChannel: NSObject, RTCPeerConnectionDelegate {
 
     signalingChannel.internalHandlers.onReceive = { [weak self] signaling in
       self?.handleSignalingOverWebSocket(signaling)
+    }
+
+    signalingChannel.internalHandlers.onReceiveJSON = { [weak self] json in
+      self?.internalHandlers.onReceiveSignalingJSON?(json)
     }
   }